% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova}
\alias{anova}
\title{Analysis of Variance}
\usage{
anova(formula, source)
}
\arguments{
\item{formula}{A formula with the form of response ~ predictors.}

\item{source}{A R data frame that contains the source data.}
}
\description{
Generate ANOVA table of a fitted model.
}
\examples{
anova(mpg ~ hp + wt, mtcars)

}
