% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_model.R
\name{lr}
\alias{lr}
\title{Linear Regression Model}
\usage{
lr(formula, source, intercept = TRUE)
}
\arguments{
\item{formula}{A formula with the form of response ~ predictors.}

\item{source}{A R data frame that contains the source data.}

\item{intercept}{An indicator for determining whether the model should be fit with or without intercept.The
default value of intercept is TRUE.}
}
\value{
A list that contains the attributes of the fitted model. All of the attributes can be visited by using $.
}
\description{
lr is a custom function designed to fit linear models. It works for both simple linear regression and multiple
linear regression and generates the same results as lm function from R.
}
\examples{
lr(mpg ~ hp + wt, mtcars)
fitted_model <- lr(mpg ~ hp + wt, mtcars)
fitted_model$r_square

}
