% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_model.R
\name{prediction}
\alias{prediction}
\title{Model Predictions}
\usage{
prediction(lr_model, new_data)
}
\arguments{
\item{lr_model}{A linear regression model generated by the lr function}

\item{new_data}{A R data frame that contains}
}
\description{
Predict the value based on the input data and trained linear regression model.
}
\examples{
fitted_model <- lr(mpg ~ hp + wt, mtcars)
new_data <- data.frame(hp = c(110), wt = c(3.521)
prediction(fitted_model, new_data)

}
