% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_model.R
\name{summary_lr}
\alias{summary_lr}
\title{Display the summary of a lr model}
\usage{
summary_lr(lr_model)
}
\arguments{
\item{lr_model}{A linear regression model generated by the lr function}
}
\description{
summary_lr function organizes and outputs each attribute of a lr model in the similar format as summary function
in R.
}
\examples{
fitted_model <- lr(mpg ~ hp + wt, mtcars)
summary_lr(fitted_model)

}
