% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_model.R
\name{prediction}
\alias{prediction}
\title{Model Prediction}
\usage{
prediction(lr_model, new_data)
}
\arguments{
\item{lr_model}{a linear regression model generated by \emph{lr} function}

\item{new_data}{a data frame containing all the variables in the model}
}
\value{
Returns the predicted values
}
\description{
Predict the value based on the input data and a trained lr model
}
\examples{
fitted_model <- lr(mpg ~ hp + wt, mtcars)
new_data <- data.frame(hp = c(110), wt = c(3.521))
p <- prediction(fitted_model, new_data)

}
