% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_model.R
\name{summary_lr}
\alias{summary_lr}
\title{Display the summary of a lr model}
\usage{
summary_lr(lr_model)
}
\arguments{
\item{lr_model}{a linear regression model generated by \emph{lr} function}
}
\description{
\emph{summary_lr} function organizes and outputs all attributes of a lr model in the similar format as \emph{summary} function
does.
}
\examples{
fitted_model <- lr(mpg ~ hp + wt, mtcars)
summary_lr(fitted_model)

}
